Fixed compiler warnings

+++ src/ThunderEgg/LocalData.h
@@ -88,7 +88,7 @@ template <int D> class LocalData
 		std::array<Side<D>, D - M> sides         = f.getSides();
 		int                        lengths_index = 0;
 		int                        sides_index   = 0;
-		for (int axis = 0; axis < D; axis++) {
+		for (size_t axis = 0; axis < D; axis++) {
 			if (sides[sides_index].getAxisIndex() == axis) {
 				if (sides[sides_index].isLowerOnAxis()) {
 					new_data += offset[sides_index] * strides[axis];
@@ -243,7 +243,7 @@ template <int D> class LocalData
 		std::array<Side<D>, D - M> sides         = f.getSides();
 		int                        lengths_index = 0;
 		int                        sides_index   = 0;
-		for (int axis = 0; axis < D; axis++) {
+		for (size_t axis = 0; axis < D; axis++) {
 			if (sides[sides_index].getAxisIndex() == axis) {
 				if (sides[sides_index].isLowerOnAxis()) {
 					new_data += (-1 - offset[sides_index]) * strides[axis];
