+++ test/Domain_MPI2.cpp
@@ -105,8 +105,8 @@ TEST_CASE("Domain<3> local indexes match position in pinfo vector",
 	auto            domain = domain_reader.getFinerDomain();
 
 	auto pinfo_vector = domain->getPatchInfoVector();
-	for (int i = 0; i < pinfo_vector.size(); i++) {
-		CHECK(pinfo_vector[i].local_index == i);
+	for (size_t i = 0; i < pinfo_vector.size(); i++) {
+		CHECK(pinfo_vector[i].local_index == (int) i);
 	}
 }
 TEST_CASE("Schur::InterfaceDomain<3> local indexes in neighbor info are consistent",
@@ -138,7 +138,7 @@ TEST_CASE("Schur::InterfaceDomain<3> local indexes in neighbor info are consiste
 					checkIdAndLocalIndex(info.id, info.local_index);
 				} else if (type == NbrType::Fine) {
 					const FineNbrInfo<2> &info = pinfo.getFineNbrInfo(s);
-					for (int i = 0; i < info.ids.size(); i++) {
+					for (size_t i = 0; i < info.ids.size(); i++) {
 						checkIdAndLocalIndex(info.ids[i], info.local_indexes[i]);
 					}
 				} else if (type == NbrType::Coarse) {
@@ -155,7 +155,7 @@ TEST_CASE("Schur::InterfaceDomain<3> local indexes in neighbor info are consiste
 					checkIdAndLocalIndex(info.id, info.local_index);
 				} else if (type == NbrType::Fine) {
 					const FineNbrInfo<1> &info = pinfo.getFineNbrInfo(e);
-					for (int i = 0; i < info.ids.size(); i++) {
+					for (size_t i = 0; i < info.ids.size(); i++) {
 						checkIdAndLocalIndex(info.ids[i], info.local_indexes[i]);
 					}
 				} else if (type == NbrType::Coarse) {
@@ -172,7 +172,7 @@ TEST_CASE("Schur::InterfaceDomain<3> local indexes in neighbor info are consiste
 					checkIdAndLocalIndex(info.id, info.local_index);
 				} else if (type == NbrType::Fine) {
 					const FineNbrInfo<0> &info = pinfo.getFineNbrInfo(c);
-					for (int i = 0; i < info.ids.size(); i++) {
+					for (size_t i = 0; i < info.ids.size(); i++) {
 						checkIdAndLocalIndex(info.ids[i], info.local_indexes[i]);
 					}
 				} else if (type == NbrType::Coarse) {
@@ -198,8 +198,8 @@ TEST_CASE("Domain<3> global indexes match position in pinfo vector",
 		start_i = 14;
 	}
 	auto pinfo_vector = domain->getPatchInfoVector();
-	for (int i = 0; i < pinfo_vector.size(); i++) {
-		CHECK(pinfo_vector[i].global_index == start_i + i);
+	for (size_t i = 0; i < pinfo_vector.size(); i++) {
+		CHECK(pinfo_vector[i].global_index == start_i + (int) i);
 	}
 }
 TEST_CASE("Domain<3> global indexes in neighbor info are consistent",
@@ -218,7 +218,7 @@ TEST_CASE("Domain<3> global indexes in neighbor info are consistent",
 	MPI_Allreduce(global_index_to_id_in.data(), global_index_to_id_out.data(), global_index_to_id_in.size(), MPI_INT, MPI_SUM, MPI_COMM_WORLD);
 
 	map<int, int> id_to_global_index_map;
-	for (int i = 0; i < global_index_to_id_out.size(); i++) {
+	for (size_t i = 0; i < global_index_to_id_out.size(); i++) {
 		id_to_global_index_map[global_index_to_id_out[i]] = i;
 	}
 
@@ -234,7 +234,7 @@ TEST_CASE("Domain<3> global indexes in neighbor info are consistent",
 					checkIdAndLocalIndex(info.id, info.global_index);
 				} else if (type == NbrType::Fine) {
 					const FineNbrInfo<2> &info = pinfo.getFineNbrInfo(s);
-					for (int i = 0; i < info.ids.size(); i++) {
+					for (size_t i = 0; i < info.ids.size(); i++) {
 						checkIdAndLocalIndex(info.ids[i], info.global_indexes[i]);
 					}
 				} else if (type == NbrType::Coarse) {
@@ -251,7 +251,7 @@ TEST_CASE("Domain<3> global indexes in neighbor info are consistent",
 					checkIdAndLocalIndex(info.id, info.global_index);
 				} else if (type == NbrType::Fine) {
 					const FineNbrInfo<1> &info = pinfo.getFineNbrInfo(e);
-					for (int i = 0; i < info.ids.size(); i++) {
+					for (size_t i = 0; i < info.ids.size(); i++) {
 						checkIdAndLocalIndex(info.ids[i], info.global_indexes[i]);
 					}
 				} else if (type == NbrType::Coarse) {
@@ -268,7 +268,7 @@ TEST_CASE("Domain<3> global indexes in neighbor info are consistent",
 					checkIdAndLocalIndex(info.id, info.global_index);
 				} else if (type == NbrType::Fine) {
 					const FineNbrInfo<0> &info = pinfo.getFineNbrInfo(c);
-					for (int i = 0; i < info.ids.size(); i++) {
+					for (size_t i = 0; i < info.ids.size(); i++) {
 						checkIdAndLocalIndex(info.ids[i], info.global_indexes[i]);
 					}
 				} else if (type == NbrType::Coarse) {
